requirements.txt
streamlit
langchain_community
langchain_ollama
langchain_pinecone
pinecone-client
pandas
python-dotenv
-e .


